requirements.txt
django
requests
python-dotenv
moralis
django-cors-headers